--- a/other_hot_results.txt
+++ b/other_hot_results.txt
@@ -1,0 +1,2 @@
+PASS
+ok  	github.com/LynnColeArt/guda	0.002s