--- a/scripts/bench_cold.sh
+++ b/scripts/bench_cold.sh
@@ -10,6 +10,11 @@ GREEN='\033[0;32m'
 YELLOW='\033[1;33m'
 NC='\033[0m' # No Color
 
+# Change to project root directory
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+PROJECT_ROOT="$(dirname "$SCRIPT_DIR")"
+cd "$PROJECT_ROOT"
+
 # Check if running as root
 if [ "$EUID" -ne 0 ]; then 
     echo -e "${RED}This script requires sudo privileges for cache flushing${NC}"