… BENCHMARK_RESULTS.md
… COLD_CACHE_ANALYSIS.md
… axpy_hot_results.txt
… benchmark_results/all_cold_20250809_000037.txt
… benchmark_results/axpy_cold_20250809_000037.txt
… benchmark_results/dot_cold_20250809_000037.txt
… benchmark_results/fusion_cold_20250809_000037.txt
… benchmark_results/gemm_cold_20250809_000037.txt
… benchmark_results/memory_cold_20250809_000037.txt
… benchmark_test.go
… docs/BENCHMARKING_GUIDE.md
… examples/perf_validation/main.go
… gemm_hot_results.txt
… scripts/benchmark_results/all_cold_20250808_235814.txt
… scripts/benchmark_results/axpy_cold_20250808_235814.txt
… scripts/benchmark_results/dot_cold_20250808_235814.txt
… scripts/benchmark_results/fusion_cold_20250808_235814.txt
… scripts/benchmark_results/gemm_cold_20250808_235814.txt
… scripts/benchmark_results/memory_cold_20250808_235814.txt
… scripts/compare_benchmarks.py